+++ nodes/Imap/operations/email/OperationsList.ts
@@ -5,6 +5,7 @@ import { createDraftOperation } from './functions/EmailCreateDraft';
 import { downloadOperation } from './functions/EmailDownload';
 import { downloadAttachmentOperation } from './functions/EmailDownloadAttachment';
 import { getEmailsListOperation } from './functions/EmailGetList';
+import { getSingleEmailOperation } from './functions/EmailGetSingle';
 import { moveEmailOperation } from './functions/EmailMove';
 import { setEmailFlagsOperation } from './functions/EmailSetFlags';
 import { manageEmailLabelsOperation } from './functions/EmailManageLabels';
@@ -13,6 +14,7 @@ export const emailResourceDefinitions: IResourceDef = {
 	resource: resourceEmail,
 	operationDefs: [
 		getEmailsListOperation,
+		getSingleEmailOperation,
 		downloadOperation,
 		downloadAttachmentOperation,
 		moveEmailOperation,